Convert RadioItem to Form Field type radio item
